+++ front/app/api/graph_data_units/responseTypes/index.ts
@@ -1,7 +1,5 @@
 import { IIdeaImageData } from 'api/idea_images/types';
 import { IIdeaData } from 'api/ideas/types';
-import { IPhaseData } from 'api/phases/types';
-import { IProjectData } from 'api/projects/types';
 
 import {
   CommentsCountRow,
@@ -15,18 +13,6 @@ import {
   ReactionsCountRow,
   TimeSeriesResponseRow as ReactionsTimeSeriesResponseRow,
 } from 'components/admin/GraphCards/ReactionsByTimeCard/useReactionsByTime/typings';
-
-export type MostReactedIdeasResponse = {
-  data: {
-    type: 'report_builder_data_units';
-    attributes: {
-      ideas: IIdeaData[];
-      project: IProjectData;
-      phase: IPhaseData;
-      idea_images: Record<string, IIdeaImageData[]>;
-    };
-  };
-};
 
 export type SingleIdeaResponse = {
   data: {
